requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0
